run prettier

+++ browser/Collection.js
@@ -1,14 +1,13 @@
-const modify = require('modifyjs');
-const sift = require('sift').default;
+const modify = require("modifyjs");
+const sift = require("sift").default;
 
-const Cursor = require('./Cursor').default;
-const ChangeStream = require('./ChangeStream').default;
+const Cursor = require("./Cursor").default;
+const ChangeStream = require("./ChangeStream").default;
 
-const { debug, randomId } = require('./utils');
+const { debug, randomId } = require("./utils");
 const ObjectID = require("bson-objectid");
 
 class Collection {
-
   constructor(db, name, opts = {}) {
     this.db = db;
     this.name = name;
@@ -18,25 +17,20 @@ class Collection {
     this.isLocalCollection = opts.isLocalCollection || false;
     this.idType = opts.idType || "ObjectID";
   }
-  
+
   insertMissingId(doc) {
-    if (doc._id)
-      return;
-    else if (this.idType === "random")
-      doc._id = Collection.randomId();
-    else if (this.idType === 'ObjectID') {
+    if (doc._id) return;
+    else if (this.idType === "random") doc._id = Collection.randomId();
+    else if (this.idType === "ObjectID") {
       doc._id = ObjectID().toHexString();
-      if (!doc.__ObjectIDs)
-        doc.__ObjectIDs = [ '_id' ];
-      else
-        doc.__ObjectIDs.push('_id');
+      if (!doc.__ObjectIDs) doc.__ObjectIDs = ["_id"];
+      else doc.__ObjectIDs.push("_id");
     }
   }
 
   // Will be called by a collection if an update occured
   _didUpdate() {
-    if (this._didUpdateTimeout)
-      clearTimeout(this._didUpdateTimeout);
+    if (this._didUpdateTimeout) clearTimeout(this._didUpdateTimeout);
 
     this._didUpdateTimeout = setTimeout(() => this._updatesFinished(), 50);
   }
@@ -56,8 +50,7 @@ class Collection {
 
   shouldPersist(doc) {
     for (let query of this.persists) {
-      if (query(doc))
-        return true;
+      if (query(doc)) return true;
     }
     return false;
   }
@@ -67,21 +60,24 @@ class Collection {
   watch() {
     const cs = new ChangeStream(this);
     this.changeStreams.push(cs);
-    cs.on('close', () => this.changeStreams = this.changeStreams.filter(x => x !== cs));
+    cs.on(
+      "close",
+      () => (this.changeStreams = this.changeStreams.filter((x) => x !== cs))
+    );
     return cs;
   }
 
   csExec(type, data) {
     // Note, cs.exec catches errors, so no need to catch here.
-    this.changeStreams.forEach(cs => cs.exec(type, data));
+    this.changeStreams.forEach((cs) => cs.exec(type, data));
   }
 
   sendChanges(operationType, _id, data) {
-    this.csExec('change', {
+    this.csExec("change", {
       operationType,
       ...data,
       ns: { db: this.db.name, coll: this.name },
-      documentKey: { _id }
+      documentKey: { _id },
     });
   }
 
@@ -98,15 +94,15 @@ class Collection {
    *   - call _funcs
    */
 
-   /**
-    * [_insert Insert "raw" document to local database, persist on match,
-    *     notify changeStreams ]
-    * @param  {object} document - document to insert that includes ._id
-    * @return {null]}          TODO
-    */
+  /**
+   * [_insert Insert "raw" document to local database, persist on match,
+   *     notify changeStreams ]
+   * @param  {object} document - document to insert that includes ._id
+   * @return {null]}          TODO
+   */
   _insert(document) {
     if (!document._id)
-      throw new Error('no doc._id ' + JSON.stringify(document));
+      throw new Error("no doc._id " + JSON.stringify(document));
 
     if (this.documents.has(document._id)) {
       // TODO, throw error.  add upsert support
@@ -114,10 +110,9 @@ class Collection {
 
     this.documents.set(document._id, document);
 
-    if (this.shouldPersist(document))
-      this.db.idb.put(this.name, document);
+    if (this.shouldPersist(document)) this.db.idb.put(this.name, document);
 
-    this.sendChanges('insert', document._id, { fullDocument: document });
+    this.sendChanges("insert", document._id, { fullDocument: document });
   }
 
   /**
@@ -128,18 +123,16 @@ class Collection {
    * @return {object} document - the inserted document (with _id)
    */
   insert(document) {
-    const docToInsert = this.isLocalCollection ? {
-      
-        ...document,
+    const docToInsert = this.isLocalCollection
+      ? {
+          ...document,
+        }
+      : {
+          ...document,
+          __pendingInsert: true,
+          __pendingSince: this.db.getTime(),
+        };
 
-    } : {
-
-        ...document,
-      __pendingInsert: true,
-      __pendingSince: this.db.getTime(),
-
-    };
-    
     this.insertMissingId(docToInsert);
 
     this._insert(docToInsert);
@@ -153,29 +146,28 @@ class Collection {
   }
 
   findOne(query) {
-    if (typeof query === 'string')
-      return this.documents.get(query) || null;
+    if (typeof query === "string") return this.documents.get(query) || null;
 
     const matches = sift(query);
     for (let [id, doc] of this.documents) {
-      if (matches(doc))
-        return doc;
+      if (matches(doc)) return doc;
     }
     return null;
   }
 
   _update(strId, newDoc) {
-    if (typeof strId !== 'string')
-      throw new Error("_update(id, ...) expects string id, not " + JSON.stringify(strId));
+    if (typeof strId !== "string")
+      throw new Error(
+        "_update(id, ...) expects string id, not " + JSON.stringify(strId)
+      );
 
     this.documents.set(strId, newDoc);
 
-    if (this.shouldPersist(newDoc))
-      this.db.idb.put(this.name, newDoc);
+    if (this.shouldPersist(newDoc)) this.db.idb.put(this.name, newDoc);
 
     // TODO should we assert strId = newDoc._id?
 
-    this.sendChanges('update', strId, {
+    this.sendChanges("update", strId, {
       // does mongo do this?
       fullDocument: newDoc,
       /*
@@ -193,14 +185,15 @@ class Collection {
     const oldDoc = this.documents.get(strId);
 
     if (!oldDoc || oldDoc.__pendingDelete)
-      throw new Error("_updateId(strId, ...) called with id with no match: " + strId);
+      throw new Error(
+        "_updateId(strId, ...) called with id with no match: " + strId
+      );
 
     const newDoc = modify(oldDoc, newDocOrChanges);
 
     // If the same documents are the same, don't mark as changed, don't sync
     // TODO, serialization
-    if (JSON.stringify(oldDoc) === JSON.stringify(newDoc))
-      return false;
+    if (JSON.stringify(oldDoc) === JSON.stringify(newDoc)) return false;
 
     // allow multiple updates
     if (!this.isLocalCollection && !newDoc.__pendingSince) {
@@ -216,12 +209,9 @@ class Collection {
   }
 
   update(idOrSelector, newDocOrChanges) {
-    if (typeof idOrSelector === 'string') {
-
+    if (typeof idOrSelector === "string") {
       return this.updateId(idOrSelector, newDocOrChanges);
-
-    } else if (idOrSelector && typeof idOrSelector === 'object') {
-
+    } else if (idOrSelector && typeof idOrSelector === "object") {
       const query = sift(idOrSelector);
 
       const updatedDocIds = [];
@@ -234,17 +224,18 @@ class Collection {
           if (this.updateId(id, newDocOrChanges)) {
             modifiedCount++;
             updatedDocIds.push(id);
-          };
+          }
         }
       return {
         matchedCount,
         modifiedCount,
         __updatedDocsIds: updatedDocIds,
       };
-
     } else {
-      throw new Error("update(id,...) expects id to be str/obj, not "
-        + JSON.stringify(idOrSelector));
+      throw new Error(
+        "update(id,...) expects id to be str/obj, not " +
+          JSON.stringify(idOrSelector)
+      );
     }
   }
 
@@ -260,8 +251,9 @@ class Collection {
 
   _insertOrReplaceOne(doc) {
     if (!doc._id)
-      throw new Error("_insertOrReplaceOne, no `_id` field in "
-        + JSON.stringify(doc));
+      throw new Error(
+        "_insertOrReplaceOne, no `_id` field in " + JSON.stringify(doc)
+      );
 
     const existing = this.documents.has(doc._id);
 
@@ -273,62 +265,48 @@ class Collection {
   }
 
   _remove(strId) {
-    if (typeof strId !== 'string')
+    if (typeof strId !== "string")
       throw new Error("_remove(strId) expects a string id");
 
     const existingDoc = this.documents.get(strId);
-    if (!existingDoc)
-      return;
+    if (!existingDoc) return;
 
     this.documents.delete(strId);
 
-    if (this.shouldPersist(existingDoc))
-      this.db.idb.delete(this.name, strId);
+    if (this.shouldPersist(existingDoc)) this.db.idb.delete(this.name, strId);
 
-    this.sendChanges('delete', existingDoc._id);
+    this.sendChanges("delete", existingDoc._id);
   }
 
   removeId(strId) {
     const doc = this.documents.get(strId);
-    if (!doc)
-      return;
+    if (!doc) return;
 
     if (this.isLocalCollection || doc.__pendingInsert) {
-
       this._remove(strId);
-
     } else {
-
       doc.__pendingDelete = true;
       doc.__pendingSince = this.db.getTime();
       this._update(doc._id, doc);
-
     }
 
     this._didUpdate();
   }
 
   remove(idOrSelector) {
-    if (typeof idOrSelector === 'string') {
-
+    if (typeof idOrSelector === "string") {
       return this.removeId(idOrSelector);
-
-    } else if (idOrSelector && typeof idOrSelector === 'object') {
-
+    } else if (idOrSelector && typeof idOrSelector === "object") {
       const query = sift(idOrSelector);
       for (let [id, doc] of this.documents) {
-        if (query(doc))
-          this.removeId(id);
+        if (query(doc)) this.removeId(id);
       }
-
     } else {
-
-      throw new Error('remove() called with invalid argument: '
-        + JSON.stringify(idOrSelector));
-
+      throw new Error(
+        "remove() called with invalid argument: " + JSON.stringify(idOrSelector)
+      );
     }
   }
-
 }
 
 Collection.randomId = randomId;
